Fix desktop build integration + JUCE-API bugs caught by real compilation
Verified the desktop preset against real JUCE 8 (configures with VST3/AU/CLAP) and
syntax-checked every adapter TU against the actual JUCE headers. Fixes the real
bugs that surfaced:

- top CMake: project() now sets VERSION (JUCE requires PROJECT_VERSION) and
  enables C (JUCE/CLAP have C sources; CMAKE_C_COMPILE_OBJECT was missing)
- workbench-app: juce::AudioSourceChannelInfo member is startSample, not the
  non-existent startSampleInDestBuffer
- parameter-view: rows held a juce::Label (Component = non-movable), so
  std::vector<Row> could not reallocate; store std::vector<unique_ptr<Row>>
- svf-effect: cast channel index to size_t (clean under JUCE's -Wsign-conversion)

All 5 adapter TUs now compile clean against JUCE; core test suite still 10/10;
portability gate green.

+++ core/effects/svf/svf-effect.h
@@ -39,13 +39,13 @@ public:
         sampleRate_ = static_cast<float>(ctx.sampleRate);
         numChannels_ = ctx.numChannels < kMaxChannels ? ctx.numChannels : kMaxChannels;
         for (int ch = 0; ch < numChannels_; ++ch)
-            filters_[ch].init(sampleRate_);
+            filters_[static_cast<std::size_t>(ch)].init(sampleRate_);
         applyAll();
     }
 
     void reset() noexcept {
         for (int ch = 0; ch < numChannels_; ++ch)
-            filters_[ch].reset();
+            filters_[static_cast<std::size_t>(ch)].reset();
         applyAll();
     }
 
@@ -54,7 +54,7 @@ public:
         const int samples = io.numSamples();
         for (int ch = 0; ch < channels; ++ch) {
             float* x = io.channel(ch);
-            SvfPrimitive& f = filters_[ch];
+            SvfPrimitive& f = filters_[static_cast<std::size_t>(ch)];
             for (int n = 0; n < samples; ++n)
                 x[n] = f.process(x[n]);
         }
@@ -110,15 +110,15 @@ private:
     void applyCutoff() noexcept {
         const float f = clampedCutoff();
         for (int ch = 0; ch < numChannels_; ++ch)
-            filters_[ch].setFreq(f);
+            filters_[static_cast<std::size_t>(ch)].setFreq(f);
     }
     void applyResonance() noexcept {
         for (int ch = 0; ch < numChannels_; ++ch)
-            filters_[ch].setRes(resonance_);
+            filters_[static_cast<std::size_t>(ch)].setRes(resonance_);
     }
     void applyMode() noexcept {
         for (int ch = 0; ch < numChannels_; ++ch)
-            filters_[ch].setMode(mode_);
+            filters_[static_cast<std::size_t>(ch)].setMode(mode_);
     }
     void applyAll() noexcept {
         applyCutoff();
